Renames internal to cmd
