+++ cmd/installer/installPkgs.go
@@ -1,0 +1,12 @@
+package installer
+
+import (
+	"errors"
+
+	"github.com/Linkinlog/gasible/cmd/osHandler"
+	"github.com/Linkinlog/gasible/internal/models"
+)
+
+func Installer(s osHandler.System, pkgManagerConf models.PackageInstallerConfig) error {
+    return errors.New("WIP")
+}
