--- internal/installer/installPkgs.go
@@ -1,12 +1,0 @@
-package installer
-
-import (
-	"errors"
-
-	"github.com/Linkinlog/gasible/internal/osHandler"
-	"github.com/Linkinlog/gasible/internal/yamlParser"
-)
-
-func Installer(s osHandler.System, pkgManagerConf yamlParser.PackageInstallerConfig) error {
-    return errors.New("WIP")
-}
